Refactored admin page into seperate components for easier readability
